Add delay in init/fini and a printf() in fini.
Add artificial delay in both initialization and finalization routines.
This is to avoid capturing output of a previously loaded program when a
new program is being loaded onto the board.

Add a printf() in the finalization routine.  This serves as a universal
output across all the programs we have, to be captured by our scripts as
a sign of a program finishing execution.