Separate the flash tool command into two commands.
Separate the flash tool command into two commands, in which the first
command loads the program image onto the board (but doesn't start
execution) and the second command resets the board (to start execution).

This is to fix cases where some program image would somehow trigger the
processor to run in non-secure mode when being loaded by the flash tool
with the --flash-load-exec argument.