Add a way to add extra settings in gen_cproject.py.
Add a way to add extra settings in gen_cproject.py that cannot be
specified statically.  The main purpose of adding extra settings is to
add LDFLAGS based on a specific program and configuration, such as
those size options for Randezvous.

This commit alse moves existing size options (all in CoreMark-Pro) to
the extra settings and adds size options for execute-in-place
configurations.